# release-2.1: opt: simplify merge and lookup join filters